refactor: move Platform and other stuff to definition layer, eliminating the need for the connect package
