--- core/connect/.gitignore
@@ -1,2 +1,0 @@
-dist/
-.coverage/
