--- core/connect/.prettierrc
@@ -1,6 +1,0 @@
-{
-  "tabWidth": 2,
-  "singleQuote": true,
-  "trailingComma": "all",
-  "semi": true
-}
