--- core/connect/eslintrc.json
@@ -1,20 +1,0 @@
-{
-  "env": {
-    "node": true
-  },
-  "root": true,
-  "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint"],
-  "extends": [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended",
-    "prettier"
-  ],
-  "rules": {
-    "comma-dangle": ["error", "always-multiline"],
-    "semi": ["error", "always"],
-    "@typescript-eslint/explicit-module-boundary-types": ["error"],
-    "@typescript-eslint/no-non-null-assertion": ["error"],
-    "@typescript-eslint/no-explicit-any": ["error", { "ignoreRestArgs": true }]
-  }
-}
