--- core/connect/jest.config.ts
@@ -1,20 +1,0 @@
-import type { JestConfigWithTsJest } from "ts-jest";
-
-const jestConfig: JestConfigWithTsJest = {
-  preset: "ts-jest",
-  verbose: true,
-  modulePathIgnorePatterns: ["mocks"],
-  moduleNameMapper: {
-    "@noble/secp256k1": require.resolve("@noble/secp256k1"),
-  },
-  transform: {
-    "^.+\\.tsx?$": [
-      "ts-jest",
-      {
-        isolatedModules: true,
-      },
-    ],
-  },
-};
-
-export default jestConfig;
