--- core/connect/package.json
@@ -1,69 +1,0 @@
-{
-  "name": "@wormhole-foundation/connect-sdk",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/wormhole-foundation/connect-sdk.git"
-  },
-  "bugs": {
-    "url": "https://github.com/wormhole-foundation/connect-sdk/issues"
-  },
-  "homepage": "https://github.com/wormhole-foundation/connect-sdk#readme",
-  "directories": {
-    "test": "__tests__"
-  },
-  "license": "Apache-2.0",
-  "version": "0.1.0-beta.1",
-  "main": "./dist/cjs/index.js",
-  "module": "./dist/esm/index.js",
-  "types": "./dist/esm/index.d.ts",
-  "author": "",
-  "description": "The core package for the Connect SDK, used in conjunction with 1 or more of the chain packages",
-  "files": [
-    "dist/**/*",
-    "src/**/*"
-  ],
-  "keywords": [
-    "wormhole",
-    "sdk",
-    "typescript",
-    "connect"
-  ],
-  "engines": {
-    "node": ">=16",
-    "pnpm": ">=7"
-  },
-  "sideEffects": false,
-  "scripts": {
-    "build:cjs": "tsc -p ./tsconfig.cjs.json",
-    "build:esm": "tsc -p ./tsconfig.esm.json",
-    "build": "npm run build:cjs && npm run build:esm",
-    "rebuild": "npm run clean && npm run build:cjs && npm run build:esm",
-    "clean": "rm -rf ./dist && rm -f ./*.tsbuildinfo",
-    "lint": "npm run prettier && eslint --fix",
-    "prettier": "prettier --write ./src",
-    "test": "jest --config ./jest.config.ts __tests__/**/*.ts",
-    "coverage": "jest --config ./jest.config.ts --coverage",
-    "docs": "typedoc"
-  },
-  "devDependencies": {
-    "@types/elliptic": "^6.4.14",
-    "@types/node": "^20.4.4",
-    "@typescript-eslint/eslint-plugin": "^5.54.1",
-    "@typescript-eslint/parser": "^5.54.1",
-    "eslint": "^8.35.0",
-    "eslint-config-prettier": "^8.7.0",
-    "lint-staged": "^13.1.2",
-    "prettier": "2.8.4",
-    "typescript": "^5.1.6"
-  },
-  "dependencies": {
-    "@ethersproject/providers": "^5.7.2",
-    "@nomad-xyz/multi-provider": "^1.1.0",
-    "@wormhole-foundation/sdk-base": "*",
-    "@wormhole-foundation/sdk-definitions": "*",
-    "axios": "^1.4.0",
-    "binary-parser": "^2.2.1",
-    "elliptic": "^6.5.4",
-    "ethers": "^5.7.2"
-  }
-}
