--- core/connect/src/index.ts
@@ -1,6 +1,0 @@
-export * from './wormhole';
-export * from './types';
-export * from './constants';
-export * from './wormholeTransfer';
-export * from './tokenTransfer';
-export * from './platformMappings';
