--- core/connect/tsconfig.esm.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.esm.json",
-  "include": ["src"],
-  "compilerOptions": {
-    "outDir": "dist/esm",
-    "rootDir": "src"
-  }
-}
