+++ core/definitions/src/index.ts
@@ -3,5 +3,9 @@ export * from "./universalAddress";
 export * from "./vaa";
 export * from "./unsignedTransaction";
 export * from "./utils";
-export * from "./tokenBridge";
-export * from "./relayer";
+export * from "./platform";
+export * from './protocolInterfaces/coreBridge';
+export * from "./protocolInterfaces/tokenBridge";
+export * from "./protocolInterfaces/relayer";
+// TODO:
+// export * from "./protocolInterfaces/tokenBridgeRelayer";
