+++ core/definitions/src/platform/constants.ts
