+++ core/definitions/src/protocolInterfaces/coreBridge/coreBridge.ts
@@ -1,0 +1,4 @@
+import { PlatformName } from "@wormhole-foundation/sdk-base";
+
+// TODO:
+export interface CoreBridge<P extends PlatformName> {}
