+++ core/definitions/src/protocolInterfaces/coreBridge/index.ts
@@ -1,0 +1,2 @@
+export * from './transfer';
+export * from './coreBridge';
