+++ core/definitions/src/protocolInterfaces/tokenBridge/index.ts
@@ -1,0 +1,2 @@
+export * from './transfer';
+export * from './tokenBridge';
